Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
+use Illuminate\Auth\Access\Gate;
+
 class QuestionsController extends Controller
 {
     /**
@@ -66,6 +68,11 @@ class QuestionsController extends Controller
     public function edit(Question $question)
     {
         //
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403,'Access denied');//or view error 
+        }
+     
         return view('questions.edit',compact('question'));
     }
 
@@ -91,6 +98,10 @@ class QuestionsController extends Controller
     public function destroy(Question $question)
     {
         // //
+        if(\Gate::denies('delete-question',$question))
+        {
+            abort(403,'Access denied');//or view error 
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success','Your question has been deleted.');
     }
